common/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "astra-mcp-common"
version = "0.1.0"
description = "Shared library for Astra MCP servers"
readme = "README.md"
authors = [
    { name = "Astra MCP Team" }
]
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0.0",
    "jsonschema>=4.0.0",
    "structlog>=23.0.0",
    "gitpython>=3.1.0",
    "pathlib2>=2.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers"
testpaths = ["tests"]
asyncio_mode = "auto"